+++ pkg/analyzer/filter_fields.go
@@ -53,6 +53,14 @@ var (
 		sampleType:     sampleTypeInt,
 	}
 
+	// FieldHostname is a query result column.
+	FieldHostname = Field{
+		querySessions:  "coalesce(nullif(t.hostname, ''), ?)", // TODO change back to just "hostname" after migration
+		queryPageViews: "t.hostname",
+		queryDirection: "ASC",
+		Name:           "hostname",
+	}
+
 	// FieldPath is a query result column.
 	FieldPath = Field{
 		querySessions:  "path",
@@ -89,6 +97,16 @@ var (
 		Name:           "entries",
 	}
 
+	// FieldEntryRate is a query result column.
+	FieldEntryRate = Field{
+		querySessions:  `toFloat64OrDefault(entries / greatest((SELECT uniq(visitor_id, session_id)%s FROM "session"%s WHERE %s), 1))`,
+		queryPageViews: `toFloat64OrDefault(entries / greatest((SELECT uniq(visitor_id, session_id)%s FROM "session"%s WHERE %s), 1))`,
+		queryImported:  `toFloat64OrDefault(entries / greatest((SELECT uniq(visitor_id, session_id)%s FROM "session"%s WHERE %s) + (SELECT sum(sessions) FROM "%s" WHERE %s), 1))`,
+		queryDirection: "DESC",
+		filterTime:     true,
+		Name:           "entry_rate",
+	}
+
 	// FieldExitPath is a query result column.
 	FieldExitPath = Field{
 		querySessions:  "exit_path",
@@ -106,6 +124,16 @@ var (
 		queryDirection: "DESC",
 		sampleType:     sampleTypeInt,
 		Name:           "exits",
+	}
+
+	// FieldExitRate is a query result column.
+	FieldExitRate = Field{
+		querySessions:  `toFloat64OrDefault(exits / greatest((SELECT uniq(visitor_id, session_id)%s FROM "session"%s WHERE %s), 1))`,
+		queryPageViews: `toFloat64OrDefault(exits / greatest((SELECT uniq(visitor_id, session_id)%s FROM "session"%s WHERE %s), 1))`,
+		queryImported:  `toFloat64OrDefault(exits / greatest((SELECT uniq(visitor_id, session_id)%s FROM "session"%s WHERE %s) + (SELECT sum(sessions) FROM "%s" WHERE %s), 1))`,
+		queryDirection: "DESC",
+		filterTime:     true,
+		Name:           "exit_rate",
 	}
 
 	// FieldVisitors is a query result column.
@@ -547,6 +575,16 @@ var (
 		queryWithFill:  "WITH FILL FROM 0 TO 60",
 		timezone:       true,
 		Name:           "minute",
+	}
+
+	// FieldWeekday is a query result column.
+	FieldWeekday = Field{
+		querySessions:  "toDayOfWeek(time, 1, '%s')",
+		queryPageViews: "toDayOfWeek(time, 1, '%s')",
+		queryDirection: "ASC",
+		queryWithFill:  "WITH FILL FROM 0 TO 7",
+		timezone:       true,
+		Name:           "weekday",
 	}
 
 	// FieldEventName is a query result column.